Move combined document html cleanup to GlossaryAutoLink.
Update all links in combined documents to jump-to links (if the link is
to information in that specific combined doc otherwise link to
other/external doc).

Add auto linking to glossary itself.

Hide link/location and anchor columns in glossary in html files.

Update glossary term generation so that glossary file names only need to
include "glossary" and not be named specifically "glossary". 

Add check so that if links have already been updated they are not
updated again.

Cleanup code and add comments.